Added File Explorer Selection of CSV File
Load from microsd button now opens file explorer window to select file. File name is then passed to LoadCSVAndPlot(). Add Exception Handling to LoadCSVAndPlot(). If The File is not a CSV, error Message appears and if the CSV is not in the proper format (Frequency, Timestamp, Data) an error message appears.