Added the fully up-to-date embedded code
This code is the most up to date code as of 1/23/25. This will be the primary code to be replaced and overwritten from now onward.

Current functionality:
Full 24-bit Resolution on single and multi-channel
Full support for unipolar and bipolar channel configuration
SD card logging (uncomment it. still working on something)
Differential voltage support

TO-DO:
Finish timescaling on SD
Variable Fs
Unit Tests for Functions
General cleanup of code. Needs to be refactored slightly for readability